[node] only loads global node modules from yarn global bin directory
Previously, a hardcoded version of node would be assumed to be installed on the
system and that node version's global bin directory would be added to $PATH.
Unfortunately, that node version would often change. Instead, this update
changes it to only assume that yarn is installed and assumes users will install
global node modules with it.